+++ lib/package.json
@@ -46,7 +46,8 @@
   "devDependencies": {
     "@parcel/bundler-library": "^2.16.0",
     "@types/node": "^24.0.6",
-    "monaco-editor": "^0.52.2"
+    "monaco-editor": "^0.52.2",
+    "udsv": "^0.7.3"
   },
   "dependencies": {
     "@streamparser/json-whatwg": "^0.0.22",
